refactor: unify effective blocked-site computation
Move effective blocklist/allowlist wildcard filtering into a shared config helper and route both runtime and CLI diagnostics through the same path to guarantee matching semantics.\n\nAdds parity regressions for dotted and punycode-form rules in runtime and CLI outputs.
